Working on JS Backend.

+++ core/src/main/java/de/mirkosertic/bytecoder/core/Bytecode5xClassParser.java
@@ -182,7 +182,7 @@ public class Bytecode5xClassParser implements BytecodeClassParser {
     }
 
     private void parseConstantPool_CONSTANT_Float(DataInput aDis, BytecodeConstantPool aConstantPool) throws IOException {
-        long theBytes = aDis.readInt();
+        int theBytes = aDis.readInt();
         aConstantPool.registerConstant(new BytecodeFloatConstant(theBytes));
     }
